requirements.txt
wxpython==4.1.0
pyserial
